Get rid of `initialOptions` in components
These should only be used in views to avoid confusing data flow
